LFG 1.0 Beta

+++ index.js
@@ -1,12 +1,16 @@
 const fs = require('node:fs');
 const path = require('node:path');
 const updateCommands = require('./deploy-commands.js')
+const assign_roles = require("./assign_roles/AssignRoles.js")
+const remove_roles = require("./assign_roles/RemoveRoles.js")
+const react_lfg_handler = require("./lfg/listener_reaction.js")
+const LFG_HANDLER = require('./lfg/listener.js')
 const { Client, Events, Collection, GatewayIntentBits, Partials } = require('discord.js');
 require('dotenv').config()
 
 const client = new Client({
 	intents: [GatewayIntentBits.Guilds, GatewayIntentBits.GuildMembers, GatewayIntentBits.GuildMessages, GatewayIntentBits.GuildMessageReactions],
-	partials: [Partials.Message, Partials.Channel, Partials.Reaction, Partials.GuildMember]
+	partials: [Partials.Message, Partials.Channel, Partials.Reaction, Partials.GuildMember, Partials.User]
 });
 
 client.commands = new Collection();
@@ -29,18 +33,35 @@ for (const file of commandFiles) {
 client.once(Events.ClientReady, c => {
 	updateCommands()
 	console.log(`Ready! Logged in as ${c.user.tag}`);
+
+	const channel = client.guilds.cache.get("799471750412369980").channels.cache.find(ch => ch.name == 'Queen\'s Court')
+
+	console.log(channel.members)
+    
 });
+
+
 
 // Log in to Discord with your client's token
 client.login(process.env.TOKEN);
 
 //Solicitamos al usuario nuevo nombre y servidor
 
+
+client.on("raw", packet=>{
+	if(packet.t=="INTERACTION_CREATE" && packet.d.data.component_type == 5){
+		packet.d.component_type = 3
+		packet.d.data.component_type = 3
+		//client.actions.InteractionCreate.handle(packet.d)
+        
+
+	}
+})
+
 //Comando para poner roles
-client.on(Events.InteractionCreate, async interaction => {
-	if (interaction.isChatInputCommand()){
-	
-		const command = interaction.client.commands.get(interaction.commandName);
+client.on(Events.InteractionCreate, async (interaction) =>{
+    if (interaction.isChatInputCommand()){
+        const command = interaction.client.commands.get(interaction.commandName);
 
 		if (!command) {
 			console.error(`No command matching ${interaction.commandName} was found.`);
@@ -53,41 +74,19 @@ client.on(Events.InteractionCreate, async interaction => {
 			console.error(error);
 			await interaction.reply({ content: 'There was an error while executing this command!', ephemeral: true });
 		}
-	}else if(interaction.isButton()){
-		if(interaction.message.components[0].components[0].data.custom_id == 'add roles in-game'){
-			await interaction.message.delete()
-		}
-	}else {
-		return;
+	}else if(interaction.isButton() && interaction.customId == 'add roles in-game' ){
+		await interaction.message.delete()
+	}else if(interaction.customId == "do_nothing"){
+		
+		await interaction.update({content:`Please make a selection.`})
+	}else if(interaction.customId.includes("lfg")){
+		LFG_HANDLER(interaction)
 	}
+    return;
 });
 
-const validateReactionName = (name) => {
-	switch(name){
-		case "Deathknight":
-			return "Death Knight"
-		case "demonhunter":
-			return "Demon Hunter"
-		case "evoker":
-			return "Evoker"
-		case "Druid":
-		case "Hunter":
-		case "Mage":
-		case "Monk":
-		case "Paladin":
-		case "Priest":
-		case "Rogue":
-		case "Shaman":
-		case "Warlock":
-		case "Warrior":
-			return name;
-		default:
-			return undefined
-	}
-}
 
-client.on(Events.MessageReactionAdd, async (reaction, user) => {
-	// When a reaction is received, check if the structure is partial
+client.on(Events.MessageReactionAdd, async (reaction, user)=>{
 	if (reaction.partial) {
 		// If the message this reaction belongs to was removed, the fetching might result in an API error which should be handled
 		try {
@@ -98,22 +97,14 @@ client.on(Events.MessageReactionAdd, async (reaction, user) => {
 			return;
 		}
 	}
-
-	const reactionName = validateReactionName(reaction._emoji.name);
-	if(!reactionName) return;
-
-	const guild = client.guilds.cache.get(reaction.message.guildId)
-	const member = guild.members.cache.get(user.id)
-	
-	const role = await guild.roles.cache.find(r => r.name == reactionName)
-	if(role)
-		member.roles.add(role)
-	
+    if(reaction.message.id == "1050973948290007080" || reaction.message.id == "1050975223484252240")
+		await assign_roles(client)(reaction, user)
+	else if(reaction.message.author.tag == "Broken Shield bot#1671" && user.tag != "Broken Shield bot#1671")
+		await react_lfg_handler(reaction, user, true)
 });
 
 
-client.on(Events.MessageReactionRemove, async (reaction, user) => {
-	// When a reaction is received, check if the structure is partial
+client.on(Events.MessageReactionRemove, async(reaction, user)=>{
 	if (reaction.partial) {
 		// If the message this reaction belongs to was removed, the fetching might result in an API error which should be handled
 		try {
@@ -124,27 +115,54 @@ client.on(Events.MessageReactionRemove, async (reaction, user) => {
 			return;
 		}
 	}
+	if(reaction.message.id == "1050973948290007080" || reaction.message.id == "1050975223484252240")
+		await remove_roles(client)(reaction, user)
+	else if(reaction.message.author.tag == "Broken Shield bot#1671" && user.tag != "Broken Shield bot#1671")
+		await react_lfg_handler(reaction, user, false)
 	
-	const reactionName = validateReactionName(reaction._emoji.name);
-	if(!reactionName) return;
-
-
-	const guild = client.guilds.cache.get(reaction.message.guildId)
-	const member = guild.members.cache.get(user.id)
-	
-	const role = await guild.roles.cache.find(r => r.name == reactionName)
-	if(role)
-		member.roles.remove(role)
-
 });
 
-client.on(Events.GuildMemberUpdate, async (oldMember, newMember)=> {
+client.on(Events.GuildMemberUpdate, require("./general_listeners/GuildMemberUpdate.js"))
+
+//process reactions
+/*
+async (oldMember, newMember)=> {
 	if( newMember.communicationDisabledUntilTimestamp != null && 
 		newMember.communicationDisabledUntilTimestamp != oldMember.communicationDisabledUntilTimestamp &&
 		newMember.user.username == "emilianohg22" &&
 		newMember.user.discriminator == "3442"
 		){
-			newMember.timeout(null)
+        	try{
+            	newMember.timeout(null);
+            }catch(err){
+                console.log("failed to remove timeout", err);
+            }
+			
 		}
-})
+}
+const addRoleBasedReaction = async (reaction, user) => {
+    if (reaction.partial) {
+		// If the message this reaction belongs to was removed, the fetching might result in an API error which should be handled
+		try {
+			await reaction.fetch();
+		} catch (error) {
+			console.error('Something went wrong when fetching the message:', error);
+			// Return as `reaction.message.author` may be undefined/null
+			return;
+		}
+	}
 
+	const reactionName = validateReactionName(reaction._emoji.name);
+	if(!reactionName) return;
+
+	const guild = client.guilds.cache.get(reaction.message.guildId)
+	const member = await guild.members.fetch(user.id)
+	
+	const role = await guild.roles.cache.find(r => r.name == reactionName)
+	if(role && member && !member.roles.cache.some(role=> role.name == reactionName)){
+		member.roles.add(role)
+        console.log("Adding role ("+reactionName+") to "+member.user.username)
+    }
+}
+
+*/
